requirements.txt
pandas
argparse
pathlib
python-dotenv
requests
tqdm